#40 made filter for ban users, also admin can't ban users with the role admin (update-user.jsp, updareUserServlet)

+++ web/src/main/java/com/belkvch/finances/financesApp/servlet/DeleteServlet.java
@@ -1,25 +1,37 @@
 package com.belkvch.finances.financesApp.servlet;
 
 import com.belkvch.finances.financesApp.dao.DefaultOperationsDAO;
+import com.belkvch.finances.financesApp.dao.DefaultUserDAO;
 import com.belkvch.finances.financesApp.entyti.Operations;
+import com.belkvch.finances.financesApp.entyti.Role;
+import com.belkvch.finances.financesApp.entyti.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet("/delete")
 public class DeleteServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        try {
-            int id = Integer.parseInt(request.getParameter("id"));
-            Operations operation = DefaultOperationsDAO.getInstance().getOperationById(id);
-            DefaultOperationsDAO.getInstance().deleteOperation(operation);
-            response.sendRedirect(request.getContextPath() + "/operations?id=" + operation.getAccountId());
-        } catch (Exception ex) {
-            response.sendRedirect("/error");
+        HttpSession httpSession = request.getSession(true);
+        int userId = (int) httpSession.getAttribute("id");
+        User currenUser = DefaultUserDAO.getInstance().getUserById(userId);
+        Role roleBan = new Role("BAN");
+        if (currenUser.getRoleId().getName().equals(roleBan.getName())) {
+            request.getRequestDispatcher("/ban.jsp").forward(request, response);
+        } else {
+            try {
+                int id = Integer.parseInt(request.getParameter("id"));
+                Operations operation = DefaultOperationsDAO.getInstance().getOperationById(id);
+                DefaultOperationsDAO.getInstance().deleteOperation(operation);
+                response.sendRedirect(request.getContextPath() + "/operations?id=" + operation.getAccountId());
+            } catch (Exception ex) {
+                response.sendRedirect("/error");
+            }
         }
     }
 }
